requirements.txt
asgiref==3.2.10
Django==3.0.8
django-autoslug-iplweb==1.9.4.dev0
django-cities==0.6
django-model-utils==4.0.0
djangorestframework==3.11.0
progressbar2==3.51.4
python-dotenv==0.14.0
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
tqdm==4.48.0
Unidecode==1.1.1
setuptools~=45.2.0